+++ lib/src/reference.dart
@@ -1,3 +1,4 @@
+/// Provides access to Git references such as branches and tags.
 import 'dart:ffi';
 
 import 'package:equatable/equatable.dart';