+++ test/remote_test.dart
@@ -261,10 +261,7 @@ void main() {
       );
     });
 
-    test(
-      "returns remote repo's reference list",
-      tags: 'remote_fetch',
-      () {
+    test("returns remote repo's reference list", tags: 'remote_fetch', () {
       Remote.setUrl(
         repo: repo,
         remote: 'libgit2',